pyproject.toml
[project]
name = "llm-course-backend"
version = "0.1.0"
description = "FastAPI + PostgreSQL backend for desktop Socratic learning platform"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.6",
  "uvicorn[standard]==0.32.1",
  "SQLAlchemy==2.0.36",
  "psycopg2-binary==2.9.10",
  "pydantic-settings==2.7.0",
  "PyJWT==2.10.1",
  "email-validator==2.2.0",
  "python-dotenv==1.0.1",
  "python-multipart==0.0.20",
  "alembic==1.14.0",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.4",
  "httpx==0.28.1",
]

[tool.uv]
package = false
